requirements.txt
streamlit
google-generativeai
pydantic
PyPDF2
pillow
pytesseract
python-dotenv
